chore(release): move release-notes tests into scripts/test/
